fun::map lazy evaluation (generator)
fun::map now has lazy evaluation (or generator) semantics, making it
similar to map in python and haskell.

This allows for a large range of techniques to be applied, but most
importantly it allows chaining of transformations without using
temporary storage, greatly adding flexibility. It also means not
everything will have to be programmed around a vector, and that it can
play directly with std::algorithms and arbitrary containers.

Carries semantics from all iterators, i.e. if some memory an iterator
depends on changes then all evaluation is undefined.